Compatible with https://esm.sh
I'm making a cloud playground website that can import any library registered in the NPM registry with `esm.sh`.

By the way, as `esm.sh` has been published for `deno`, it is not possible to import `NodeJS` built-in libraries.

In such reason, this PR removes every direct import statements to the NodeJS built-in libraries, and performed indirect import through `import2` library, so that bundler cannot merge them.